Issue-213: Moved CreditsView to separate module.
