+++ app/src/main/res/layout/layout_tv_details_content.xml
@@ -121,7 +121,7 @@
         app:layout_constraintTop_toBottomOf="@+id/vKeywords"
         tools:text="Tv Overview" />
 
-    <com.kirchhoff.movies.ui.screens.core.credits.CreditsView
+    <com.kirchhoff.movies.creditsview.CreditsView
         android:id="@+id/vCredits"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
