+++ app/src/main/res/values/strings.xml
@@ -5,8 +5,6 @@
     <string name="menu_persons">Persons</string>
     <string name="retry">Retry</string>
     <string name="trailers">Trailers</string>
-    <string name="cast_credits">Cast Credits</string>
-    <string name="crew_credits">Crew Credits</string>
 
     <string name="tv_seasons_format">Seasons: %d</string>
     <string name="tv_episodes_format">Episodes: %d</string>
