+++ creditsview/src/main/java/com/kirchhoff/movies/creditsview/adapter/CreditsAdapter.kt
@@ -1,9 +1,9 @@
-package com.kirchhoff.movies.ui.screens.core.credits.adapter
+package com.kirchhoff.movies.creditsview.adapter
 
 import android.view.ViewGroup
 import com.kirchhoff.movies.core.ui.recyclerview.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.core.ui.recyclerview.BaseVH
-import com.kirchhoff.movies.ui.screens.core.credits.CreditsView
+import com.kirchhoff.movies.creditsview.CreditsView
 
 class CreditsAdapter(itemClickListener: OnItemClickListener<CreditsView.CreditsInfo>) :
     BaseRecyclerViewAdapter<BaseVH<CreditsView.CreditsInfo>, CreditsView.CreditsInfo>(itemClickListener = itemClickListener) {
