+++ creditsview/src/main/res/layout/item_credit.xml
