+++ creditsview/src/main/res/layout/view_credits.xml
