+++ creditsview/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="cast_credits">Cast Credits</string>
+    <string name="crew_credits">Crew Credits</string>
+</resources>
